[#192] refactor: popupStoreBundle 리페키징
