+++ src/main/java/com/sparta/popupstore/domain/popupstore/bundle/dto/response/PopupStoreOperatingResponseDto.java
@@ -1,6 +1,6 @@
-package com.sparta.popupstore.domain.popupstore.dto.response;
+package com.sparta.popupstore.domain.popupstore.bundle.dto.response;
 
-import com.sparta.popupstore.domain.popupstore.entity.PopupStoreOperating;
+import com.sparta.popupstore.domain.popupstore.bundle.entity.PopupStoreOperating;
 import lombok.Getter;
 
 import java.time.DayOfWeek;
