+++ src/main/java/com/sparta/popupstore/domain/popupstore/bundle/entity/PopupStoreOperating.java
@@ -1,5 +1,6 @@
-package com.sparta.popupstore.domain.popupstore.entity;
+package com.sparta.popupstore.domain.popupstore.bundle.entity;
 
+import com.sparta.popupstore.domain.popupstore.entity.PopupStore;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
