+++ src/main/java/com/sparta/popupstore/domain/popupstore/bundle/enums/PopupStoreAttributeEnum.java
@@ -1,0 +1,7 @@
+package com.sparta.popupstore.domain.popupstore.bundle.enums;
+
+public enum PopupStoreAttributeEnum {
+    PARKING,
+    RESERVATION,
+    ;
+}
