+++ src/main/java/com/sparta/popupstore/domain/popupstore/bundle/repository/PopupStoreAttributeRepository.java
@@ -1,7 +1,7 @@
-package com.sparta.popupstore.domain.popupstore.repository;
+package com.sparta.popupstore.domain.popupstore.bundle.repository;
 
 import com.sparta.popupstore.domain.popupstore.entity.PopupStore;
-import com.sparta.popupstore.domain.popupstore.entity.PopupStoreAttribute;
+import com.sparta.popupstore.domain.popupstore.bundle.entity.PopupStoreAttribute;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
