+++ src/main/java/com/sparta/popupstore/domain/popupstore/bundle/repository/PopupStoreOperatingRepository.java
@@ -1,7 +1,7 @@
-package com.sparta.popupstore.domain.popupstore.repository;
+package com.sparta.popupstore.domain.popupstore.bundle.repository;
 
 import com.sparta.popupstore.domain.popupstore.entity.PopupStore;
-import com.sparta.popupstore.domain.popupstore.entity.PopupStoreOperating;
+import com.sparta.popupstore.domain.popupstore.bundle.entity.PopupStoreOperating;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
