--- src/main/java/com/sparta/popupstore/domain/popupstore/entity/PopupStoreAttributeEnum.java
@@ -1,7 +1,0 @@
-package com.sparta.popupstore.domain.popupstore.entity;
-
-public enum PopupStoreAttributeEnum {
-    PARKING,
-    RESERVATION
-    ;
-}
