+++ src/main/java/com/sparta/popupstore/domain/reservation/service/ReservationService.java
@@ -3,9 +3,9 @@ package com.sparta.popupstore.domain.reservation.service;
 import com.sparta.popupstore.domain.common.exception.CustomApiException;
 import com.sparta.popupstore.domain.common.exception.ErrorCode;
 import com.sparta.popupstore.domain.popupstore.entity.PopupStore;
-import com.sparta.popupstore.domain.popupstore.entity.PopupStoreAttribute;
-import com.sparta.popupstore.domain.popupstore.entity.PopupStoreAttributeEnum;
-import com.sparta.popupstore.domain.popupstore.repository.PopupStoreAttributeRepository;
+import com.sparta.popupstore.domain.popupstore.bundle.entity.PopupStoreAttribute;
+import com.sparta.popupstore.domain.popupstore.bundle.enums.PopupStoreAttributeEnum;
+import com.sparta.popupstore.domain.popupstore.bundle.repository.PopupStoreAttributeRepository;
 import com.sparta.popupstore.domain.popupstore.repository.PopupStoreRepository;
 import com.sparta.popupstore.domain.reservation.dto.request.ReservationCreateRequestDto;
 import com.sparta.popupstore.domain.reservation.entity.Reservation;
